+++ front/src/types/db/index.ts
@@ -1,1 +1,2 @@
-export type { CurrentUser, Profile, TravelinksType } from '@/types/db/user/user'
+export type { CurrentUser, Profile } from '@/types/db/user'
+export type { TravelinkRequestType } from '@/types/db/travelink'
